Move common code from data_ingestion and object_detection to common_utils
