+++ common_utils/common_utils/file_manager.py
@@ -6,6 +6,9 @@ import os
 from datetime import datetime
 from typing import Optional
 
+_PATH_SUFFIX = "_path"
+_DIR_SUFFIX = "_dir"
+
 def get_latest_image_timestamp(folder_path: str) -> Optional[datetime]:
     """
     Finds the latest image timestamp in the folder, or returns None if no images are found.
@@ -48,3 +51,40 @@ def get_latest_folder(camera_path: str) -> Optional[str]:
     """
     date_folders = os.listdir(camera_path)
     return max(date_folders) if date_folders else None
+
+
+def update_relative_paths(params, root_dir) -> None:
+    """
+    Update the relative paths in the parameters dictionary with the root directory.
+    Only keys ending with '_path' or '_dir' are updated.
+
+    Parameters
+    ----------
+    params : dict
+        The parameters dictionary.
+
+        A dictionary as follows:
+        ```
+        {
+            'key1': 'relative/path/to/file1',
+            'key2': 'relative/path/to/file2',
+            ...
+        }
+        ```
+
+    keys : List[str]
+        The keys to update.
+
+    root_dir : str
+        The root directory.
+
+    Returns
+    -------
+    params : dict
+        The updated parameters dictionary.
+    """
+    for key, path in params.items():
+        if key.endswith(_PATH_SUFFIX) or key.endswith(_DIR_SUFFIX):
+            params[key] = os.path.join(root_dir, path)
+
+    return params
