+++ data_ingestion/config.json
@@ -3,7 +3,7 @@
         "base_url": "https://images.drivebc.ca/ReplayTheDay/archive/",
         "cameras": ["147", "148", "149", "150"],
         "image_dir": "dat/raw_images/",
-        "log_file": "logs/traffic_image_downloader.log"
+        "log_file_path": "logs/traffic_image_downloader.log"
     },
     "preprocessor_params": {
         "source_dir": "dat/raw_images/",
